Create Create_Single_User
Facilitates the creation of a single AD user; you will be prompted for input on most questions.  You can use if...elseif loops in place of $OU, $domain, etc to assign your domains and OU's based on other characteristics; I use an if..else loop based on the variable $unit to determine what script they get, and what OU they are assigned to.